[makefiles] Create ~/Applications/makefiles/
This is an idea I've had to utilize Makefiles, a common and standard
tool in Unix-like operating systems, to manage installing programs.

It's in my opinion a good way to re-install everything when switching
machines without having to remember everything you've installed and how
to install it.

I'll see if I'll abandon the idea in the future, but currently it was
nice to install Clojure this way. Even if it took a bit longer.